backend/requirements.txt
fastapi
uvicorn
redis
celery
pydantic
pydantic-settings
email-validator
passlib[bcrypt]
python-multipart
httpx
playwright
google-genai
beautifulsoup4
pytest
python-dotenv
motor
beanie
PyJWT
